requirement.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
DateTime==4.3
idna==2.8
numpy==1.16.3
pandas==0.24.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.1
urllib3==1.24.3
zope.interface==4.6.0
